+++ client/app/form-builder/adv-form-builder-def.js
@@ -3,7 +3,7 @@
  */
 var advFormBuilderDef = {
   "type": "LOINC",
-  "code": "basicItemLevelFields",
+  "code": "advancedItemLevelFields",
   "name": "Advanced",
   "templateOptions": {
     "allowHTMLInInstructions": true,
@@ -63,6 +63,77 @@ var advFormBuilderDef = {
             "value": "none"
           }
         ]
+      }
+    },
+    {
+      // ********** Hidden item to indicate externallyDefined from basic widget ********,
+      "questionCode": "__itemTypeRef",
+      "question": "You should never see this!",
+      "dataType": "ST",
+      "header": false,
+      "displayControl": {
+        "css": [
+          {
+            "name": "display",
+            "value": "none"
+          }
+        ]
+      }
+    },
+    {
+      // *********** formula/calculationMethod ************************,
+      "questionCode": "_calculationMethod",
+      "question": "Calculation [2]",
+      "dataType": "CNE",
+      "header": false,
+      "answers": "_calculationMethod",
+      "codingInstructions": "Applies if siblings have answer list with scores. Contains the calculation method (formula) in human readable form, for calculating the value of any measure that is based on an algebraic or other formula.",
+      "skipLogic": {
+        "conditions": [
+          {
+            "source": "_dataType",
+            "trigger": {"value": "QTY"}
+          },
+          {
+            "source": "_dataType",
+            "trigger": {"value": "REAL"}
+          },
+          {
+            "source": "_dataType",
+            "trigger": {"value": "INT"}
+          }
+        ],
+        "action": "show",
+        "logic": "ANY"
+      },
+      "answerCardinality": {
+        "min": "0",
+        "max": "1"
+      },
+      "value": {
+        "text": "None",
+        "code": "none"
+      },
+    },
+    {
+      "questionCode": "calculatedExpression",
+      "question": "FHIRPath calculated expression [2]",
+      "dataType": "TX",
+      "header": false,
+      "codingInstructions": "Calculated value for a question answer as determined by an FHIRPath evaluated expression. The expression is not validated for syntax or semantics. It is author's responsibility to enter correct expression.",
+      "skipLogic": {
+        "conditions": [
+          {
+            "source": "__itemTypeRef",
+            "trigger": {"value": "question"}
+          },
+          {
+            "source": "_calculationMethod",
+            "trigger": {"value": {"code": "calculatedExpression"}}
+          }
+        ],
+        "action": "show",
+        "logic": "ALL"
       }
     },
     {
@@ -684,8 +755,33 @@ var advFormBuilderDef = {
       "question": "Copyright Notice",
       "dataType": "ST",
       "header": false,
-      "codingInstructions": "Add any copy right notice text you wish to include for this item."
-    }
-
+      "codingInstructions": "Add any copyright notice text you wish to include for this item."
+    },
+    {
+      "questionCode": "_fhirVariables",
+      "question": "FHIR Variable",
+      "header": true,
+      "questionCardinality": {
+        "min": "1",
+        "max": "*"
+      },
+      "items": [
+        {
+          "questionCode": "name",
+          "question": "Name*",
+          "dataType": "ST"
+        },
+        {
+          "questionCode": "expression",
+          "question": "FHIR Path expression",
+          "dataType": "TX"
+        },
+        {
+          "questionCode": "description",
+          "question": "Description",
+          "dataType": "TX"
+        }
+      ]
+    },
   ]
 };