+++ client/app/form-builder/answer-lists.js
@@ -2,7 +2,7 @@
  * Answer lists to be used in form builder. The form-builder-data.js has references here
  * to include them in form builder data model.
  *
- * @type {{minCardinality: *[], maxCardinality: *[], boolean: *[], template: *[], questionCodeSystem: *[], calculationMethod: *[], dataType: *[], editable: *[], testUnits: *[], ucumUnits: *[]}}
+ * @type {{minCardinality: *[], maxCardinality: *[], boolean: *[], template: *[], questionCodeSystem: *[], _calculationMethod: *[], dataType: *[], editable: *[], testUnits: *[], ucumUnits: *[]}}
  *
  *
  */
@@ -267,18 +267,18 @@ var answerLists = {
       "code": "Custom"
     }
   ],
-  "calculationMethod": [
+  "_calculationMethod": [
     {
       "text": "None",
       "code": "none"
     },
     {
-      "text": "TOTAL SCORE",
+      "text": "Score calculation",
       "code": "TOTALSCORE"
     },
     {
-      "text": "BMI",
-      "code": "BMI"
+      "text": "FHIRPath calculated expression",
+      "code": "calculatedExpression"
     }
   ],
   "dataType": [
@@ -321,6 +321,10 @@ var answerLists = {
     {
       "text": "Phone",
       "code": "PHONE"
+    },
+    {
+      "text": "Quantity",
+      "code": "QTY"
     },
     {
       "text": "Ratio",
@@ -392,6 +396,29 @@ var answerLists = {
       "code": "unknown",
       "text": "Unknown"
     },
+  ],
 
-  ]
+  "itemType": [
+    {
+      "code": "question",
+      "text": "Question"
+    },
+    {
+      "code": "group",
+      "text": "Group"
+    },
+    {
+      "code": "display",
+      "text": "Display"
+    }/*,
+    {
+      "code": "TOTALSCORE",
+      "text": "TOTAL SCORE"
+    },
+    {
+      "code": "calculatedExpression",
+      "text": "FHIRPath calculated expression"
+    }
+    */
+  ],
 };