+++ client/app/form-builder/basic-form-builder-def.js
@@ -17,17 +17,67 @@ var formBuilderDef = {
   },
   "items": [
     {
+      /*
+      ******** item type: 'group' determines item.header=true, 'display' determines item.dataType=TITLE, and
+      * 'question' will show dataType pull down to get other item.dataType values
+      */
+      "questionCode": "__itemType",
+      "question": "Item type*",
+      "dataType": "CNE",
+      "header": false,
+      "answers": "itemType",
+      "codingInstructions": "Choose the type item you want to create",
+      "answerCardinality": {
+        "min": "1",
+        "max": "1"
+      },
+      "value": {
+        "code": "question",
+        "text": "Question"
+      }
+    },
+    {
+      // *********** Data type ************************,
+      "questionCode": "dataType",
+      "question": "Data type*",
+      "dataType": "CNE",
+      "header": false,
+      "answers": "dataType",
+      "codingInstructions": "Enter the data type of the answer. Valid data types are:",
+      "displayControl": {
+        "answerLayout": {
+          "type": "COMBO_BOX"
+        }
+      },
+      "answerCardinality": {
+        "min": "1",
+        "max": "1"
+      },
+      "value": {
+        "text": "String",
+        "code": "ST"
+      },
+      "skipLogic": {
+        "conditions": [
+          {
+            "source": "__itemType",
+            "trigger": {"value": {"code": "question"}}
+          }
+        ],
+        "action": "show"
+      }
+    },
+    {
       // *********** question ************************,
       "questionCode": "question",
-      "question": "Text*",
+      "question": "Item name*",
       "dataType": "ST",
       "header": false,
-      //"codingInstructions": "Enter wording for question."
       "codingInstructions": "This is required: Enter the section header or question text exactly as it is displayed on your form."
     },
     {
       "questionCode": "questionCodeSystem",
-      "question": "Coding system",
+      "question": "Item coding system",
       "dataType": "CNE",
       "answers": "questionCodeSystem",
       "header": false,
@@ -45,7 +95,7 @@ var formBuilderDef = {
     {
       // *********** questionCode ************************,
       "questionCode": "questionCode",
-      "question": "Code*",
+      "question": "Item code*",
       "dataType": "ST",
       "header": false,
       "answerCardinality": {
@@ -53,7 +103,6 @@ var formBuilderDef = {
         "max": "1"
       },
       "editable": "1",
-      // "codingInstructions": "Enter a code that matches the coding system you selected; or, create your own unique code."
       "codingInstructions": "This is required: Enter the unique question code for the question or section header given in the Text field. <p>If a question or section header is not available, enter any unique identifier in square brackets, e.g., [Q1], [Q2], [H1].</p>"
     },
     {
@@ -75,10 +124,9 @@ var formBuilderDef = {
     {
       // *********** codingInstructions ************************,
       "questionCode": "codingInstructions",
-      "question": "Question instructions [1]",
+      "question": "Help text for the item [1]",
       "dataType": "ST",
       "header": false,
-      //"codingInstructions": "Enter any explanatory text needed to help the user answer the question, such as \"Select all that apply\". Instructions will appear before the question."
       "codingInstructions": "Instructions for the person completing the form on how to answer a specific item. This could include additional explanatory text that supplements the question or the number of expected responses."
     },
     // ************* repeatQuestion ***************************,
@@ -95,21 +143,24 @@ var formBuilderDef = {
       }
     },
     {
-      /*
-      ******** header (hidden from the user with a dummy source skip logic) **
-      * There is no mechanism to hide an item in lforms.
-      * Picking a dummy non existent source is not an option after  lforms#5.0 version.
-      * Instead use a source whose value never satisfies skip logic of this item.
-      */
-      "questionCode": "header",
-      "question": "Section",
-      "dataType": "CNE",
+      "questionCode": "answerRequired",
+      "question": "Answer required?",
+      "codingInstructions": "Choose 'Yes' to allow selection of multiple answers from the the answer list.",
       "header": false,
+      "dataType": "CNE",
       "answers": "boolean",
-      "codingInstructions": "If you choose 'Yes', this question is used as section header",
       "value": {
         "text": "No",
         "code": false
+      },
+      "skipLogic": {
+        "conditions": [
+          {
+            "source": "__itemType",
+            "trigger": {"value": {"code": "question"}}
+          }
+        ],
+        "action": "show"
       }
     },
     {
@@ -126,8 +177,8 @@ var formBuilderDef = {
       "skipLogic": {
         "conditions": [
           {
-            "source": "header",
-            "trigger": {"value": {"code": false}}
+            "source": "__itemType",
+            "trigger": {"value": {"code": "question"}}
           }
         ],
         "action": "show"
@@ -141,54 +192,6 @@ var formBuilderDef = {
       "value": {
         "text": "Editable",
         "code": "1"
-      }
-    },
-    {
-      "questionCode": "answerRequired",
-      "question": "Answer required?",
-      "codingInstructions": "Choose 'Yes' to allow selection of multiple answers from the the answer list.",
-      "header": false,
-      "dataType": "CNE",
-      "answers": "boolean",
-      "value": {
-        "text": "No",
-        "code": false
-      },
-      "skipLogic": {
-        "conditions": [
-          {
-            "source": "header",
-            "trigger": {"value": {"code": false}}
-          }
-        ],
-        "action": "show"
-      }
-    },
-    {
-      // *********** Data type ************************,
-      "questionCode": "dataType",
-      "question": "Type",
-      "dataType": "CNE",
-      "header": false,
-      "answers": "dataType",
-      "codingInstructions": "Enter the data type of the answer. Valid data types are:",
-      "displayControl": {
-        "answerLayout": {
-          "type": "COMBO_BOX"
-        }
-      },
-      "value": {
-        "text": "String",
-        "code": "ST"
-      },
-      "skipLogic": {
-        "conditions": [
-          {
-            "source": "header",
-            "trigger": {"value": {"code": false}}
-          }
-        ],
-        "action": "show"
       }
     },
     {
@@ -330,8 +333,8 @@ var formBuilderDef = {
       "skipLogic": {
         "conditions": [
           {
-            "source": "header",
-            "trigger": {"value": {"code": false}}
+            "source": "__itemType",
+            "trigger": {"value": {"code": "question"}}
           }
         ],
         "action": "show"
@@ -384,7 +387,6 @@ var formBuilderDef = {
         "min": "0",
         "max": "*"
       },
-      //"codingInstructions": "If using an externally defined list of answers to the question, enter the url here.",
       "codingInstructions": "Example units (e.g., #/day) are provided in a dropdown list in each field. You can use one of these if appropriate or enter other units. Units of measure are not necessary for terms with fixed answer lists or free text answers.",
       "skipLogic": {
         "logic": "ANY",
@@ -424,33 +426,5 @@ var formBuilderDef = {
         "listColHeaders": ["Unit", "Name", "Guidance"]
       }
     },
-    {
-      // *********** formula/calculationMethod ************************,
-      "questionCode": "calculationMethod",
-      "question": "Score calculation method [2]",
-      "dataType": "CNE",
-      "header": false,
-      "answers": "calculationMethod",
-      //"codingInstructions": "Select one of the formulas from the list.",
-      "codingInstructions": "Applies if siblings have answer list with scores. Contains the calculation method (formula) in human readable form, for calculating the value of any measure that is based on an algebraic or other formula.",
-      "skipLogic": {
-        "conditions": [
-          {
-            "source": "header",
-            "trigger": {"value": {"code": false}}
-          }
-        ],
-        "action": "show"
-      },
-      "answerCardinality": {
-        "min": "0",
-        "max": "1"
-      },
-      "value": {
-        "text": "None",
-        "code": "none"
-      },
-
-    }
   ]
 };