+++ client/app/form-builder/formlevel-panel/adv-form-level-fields-def.js
@@ -1,6 +1,6 @@
 var advFormLevelFieldsDef = {
   "code": "advFormLevelFields",
-  "name": "Advanced Form Level Properties",
+  "name": "Form Attributes - Advanced",
   "templateOptions": {
     "allowHTMLInInstructions": true,
     "hideFormControls": true,
@@ -273,6 +273,50 @@ var advFormLevelFieldsDef = {
           "answers": "boolean",
           "header": false,
           "codingInstructions": "If this coding was chosen directly by the user.<br/><b>FHIR type: boolean</b>",
+          "codingInstructionsFormat": "html",
+        },
+      ]
+    },
+    {
+      "questionCode": "_fhirVariables",
+      "localQuestionCode": "Expression",
+      "question": "FHIRPath Variable",
+      "header": true,
+      "codingInstructions": "FHIRPath expression.<br/><b>FHIR type: Expression</b>",
+      "codingInstructionsFormat": "html",
+      "questionCardinality": {
+        "min": "0",
+        "max": "*"
+      },
+      "items": [
+        {
+          "questionCode": "name",
+          "localQuestionCode": "id",
+          "question": "Name*",
+          "dataType": "ST",
+          "header": false,
+          "codingInstructions": "Name of the questionnaire-variable.<br/><b>FHIR type: id</b>",
+          "codingInstructionsFormat": "html",
+          "restrictions": {
+            "pattern": "/^[A-Za-z0-9\\-\\.]{1,64}$/"
+          },
+        },
+        {
+          "questionCode": "expression",
+          "localQuestionCode": "string",
+          "question": "FHIRPath Expression",
+          "dataType": "TX",
+          "header": false,
+          "codingInstructions": "No syntax or semantic checks are performed. It is assumed the expression is correct. Please verify in the preview whether it is working as expected.<br/><b>FHIR type: string</b>",
+          "codingInstructionsFormat": "html",
+        },
+        {
+          "questionCode": "description",
+          "localQuestionCode": "string",
+          "question": "Description",
+          "dataType": "TX",
+          "header": false,
+          "codingInstructions": "A brief, natural language description of the condition that effectively communicates the intended semantics.<br/><b>FHIR type: string</b>",
           "codingInstructionsFormat": "html",
         },
       ]