+++ client/app/form-builder/formlevel-panel/basic-form-level-fields-def.js
@@ -1,6 +1,6 @@
 var basicFormLevelFieldsDef = {
   "code": "basicFormLevelFields",
-  "name": "Basic Form Level Properties",
+  "name": "Form Attributes - Basic",
   "templateOptions": {
     "allowHTMLInInstructions": true,
     "hideFormControls": true,