+++ test/e2e/formbuilder/formbuilder.po.js
@@ -13,7 +13,8 @@ var jsonpath = require('jsonpath');
  */
 var FormBuilder = function () {
   this.autoCompListItems = element.all(by.css('#searchResults li'));
-  this.questionTree = element(by.css('.angular-ui-tree > ol > li'));
+  //this.questionTree = element(by.css('.angular-ui-tree > ol > li'));
+  this.questionTree = element(by.id('tree-container-id'));
   this.addButton = element(by.id('sidebar-container')).element(by.css('button'));
   this.dialog = element.all(by.css('md-dialog')).last();
   this.importLOINCRadio = this.dialog.element(by.cssContainingText('md-radio-button', 'Import from LOINC'));
@@ -41,8 +42,11 @@ var FormBuilder = function () {
   this.questionAnswerText = element(by.id('/answers/text/1/1'));
   this.prefix = element(by.id('/prefix/1'));
 
-  this.basicEditTab = element(by.cssContainingText('md-tab-item', 'Build basic properties'));
-  this.advancedEditTab = element(by.cssContainingText('md-tab-item', 'Build advanced properties'));
+  this.basicEditTab = element(by.cssContainingText('md-tab-item', 'Item attributes - Basic'));
+  this.advancedEditTab = element(by.cssContainingText('md-tab-item', 'Item attributes - Advanced'));
+
+  this.basicFormEditTab = element(by.cssContainingText('md-tab-item', 'Form attributes - Basic'));
+  this.advancedFormEditTab = element(by.cssContainingText('md-tab-item', 'Form Attributes - Advanced'));
 
   this.useRestrictionsYes = element(by.id('/useRestrictions/1true'));
   this.restrictionName1 = element(by.id('/useRestrictions/restrictions/name/1/1/1'));
@@ -260,6 +264,20 @@ var FormBuilder = function () {
     for (var i = 0; i < repetitions; i++) {
       elem.sendKeys(stringOfKeys);
     }
+  };
+
+
+  /**
+   * Assign string to value property of elem. Typically used for input elements, where elem.sendKeys() messes up complex
+   * strings.
+   *
+   * @param {Object} elem - Such as textarea, or text input element
+   * @param {String} str - Input string
+   */
+  this.sendString = function(elem, str) {
+    this.scrollIntoView(elem);
+    browser.executeScript('arguments[0].value = "'+str.slice(0,-1)+'"', elem.getWebElement());
+    elem.sendKeys(str.slice(-1));
   };